Expose StringRoute
