+++ ema-examples/src/Ema/Example/Ex03_Store.hs
@@ -18,9 +18,8 @@ import Ema
 import Ema.Example.Common (tailwindLayout, watchDirForked)
 import Ema.Route.Generic
 import Ema.Route.Generic.TH (deriveGeneric, deriveIsRoute)
-import Ema.Route.Prism
+import Ema.Route.Lib.Extra.StringRoute (StringRoute (StringRoute))
 import Generics.SOP qualified as SOP
-import Optics.Core (coercedTo, iso, prism', (%))
 import System.FSNotify qualified as FSNotify
 import System.FilePath (takeFileName, (</>))
 import Text.Blaze.Html5 ((!))
@@ -75,24 +74,6 @@ data CategoryRoute
                  ]
              ]
         )
-
--- | A route represented by a stringy type; associated with a foldable of the same as its model.
-newtype StringRoute (a :: Type) r = StringRoute {unStringRoute :: r}
-  deriving stock (Show, Eq, Ord, Generic)
-  deriving anyclass (SOP.Generic, SOP.HasDatatypeInfo)
-
-instance (IsString r, ToString r, Eq r, Ord r) => IsRoute (StringRoute a r) where
-  type RouteModel (StringRoute a r) = Map r a
-  routePrism as =
-    toPrism_ $
-      htmlSuffixPrism
-        % iso fromString toString
-        % mapMemberPrism as
-        % coercedTo
-    where
-      mapMemberPrism m =
-        prism' id $ \r -> do pure r <* (guard $ r `Map.member` m)
-  routeUniverse as = StringRoute <$> Map.keys as
 
 deriveGeneric ''ProductRoute
 deriveIsRoute
