+++ ema-extra/ema-extra.cabal
@@ -100,6 +100,7 @@ library
     Ema.Route.Lib.Extra.PandocRoute
     Ema.Route.Lib.Extra.SlugRoute
     Ema.Route.Lib.Extra.StaticRoute
+    Ema.Route.Lib.Extra.StringRoute
 
   hs-source-dirs:   src
   default-language: Haskell2010
